backend/requirements.txt
Flask==1.0.2
google-cloud-datastore==1.15.3
google-cloud-storage==1.32.0
google-cloud-vision==1.0.0

